issue #19456: lock and release working

+++ source/xm/javascript/model.sql
@@ -39,11 +39,20 @@ select xt.install_js('XM','Model','xtuple', $$
   }
 
 
-  XM.Model.releaseLock = function (tableName, recordId, username) {
-    var tableOid = XT.Data.getTableOid(tableName),
+  XM.Model.releaseLock = function (recordType, recordId, username) {
+    if (!XT.Data.doLockRecords()) {
+      return false;
+    }
+    var nameSpace = recordType.beforeDot(), 
+      type = recordType.afterDot(),
+      map = XT.Orm.fetch(nameSpace, type),
+      tableOid = XT.Data.getTableOid(map.table),
       sql = "delete from xt.lock where lock_table_oid = $1 and lock_record_id = $2 and lock_username = $3";
   
-    /* if (DEBUG) */ plv8.elog(NOTICE, "Releasing lock", tableOid, recordId, username); 
+    /* make sure the id is an integer */
+    recordId = XT.Data.idAsInt(recordId);
+
+    if (DEBUG) plv8.elog(NOTICE, "Releasing lock", recordType, tableOid, recordId, username); 
     plv8.execute(sql, [tableOid, recordId, username]);
     /* TODO: should we return the number of rows affected? */
